requirements.txt
# Core utilities
dill==0.3.5.1
from-root==1.0.2
gdown
flask
flask-cors

# Jupyter (STABLE)
notebook>=6.5.5
jupyterlab>=3.6

# YOLOv5 / Computer Vision Base
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.23.5
opencv-python>=4.5
Pillow>=10.3.0
psutil
PyYAML>=5.3.1
requests>=2.32.2
scipy>=1.7
tqdm>=4.66.3

# Deep Learning (CPU versions for Windows)
torch==1.13.1+cpu
torchvision==0.14.1+cpu
ultralytics>=8.2.64
thop>=0.1.1

# Logging & Visualization
tensorboard>=2.4.1
pandas>=1.3
seaborn>=0.11

# Interactive
ipython

# Build tools
setuptools>=70.0.0
packaging

# Extras
urllib3>=2.5.0 ; python_version > "3.8"
